starts work with api

+++ core/models.py
@@ -3,7 +3,7 @@ from django.db import models
 
 class Account(models.Model):
     name = models.CharField(max_length=50)
-    credit_card = models.IntegerField()
+    credit_card = models.BigIntegerField()
     pin_code = models.IntegerField()
     balance = models.DecimalField(max_digits=10, decimal_places=2)
